Good Shape
The main robot code is in pretty good shape now. What worries me is the field simulator; what's wrong with the process such that when the robot finishes rotating, there has to be a little additional distance added to the robot's position? REALLY need to figure out the math of this.