Important Changes
Starting points of autonomous made static variables; Elevator sensor moved to the subsystem; autoTarget chooser added so that our autonomous could focus on one target at one time. Strategy-wise, this is desirable because we want to know definitely which target the robot is going for before the sides of the switches and scales are even assigned.