Auto Change
Variablized even more points. In addition to the three starting points, which we may want to change constantly, the delivery/destination points for the switches and scale should be variablized as well. Points such as the front of the left side of the scale, or the side of the left side of the scale are all locally stored, and are used in the making of auto routes when the robot initiates itself. This way, it's a lot more clear and flexible. We do not want to be frantically going through the code and changing multiple points during the competition! Yah!!!